fix: refactor common.rs into /processing
