+++ crates/server/src/handlers/blocks/mod.rs
@@ -3,6 +3,7 @@ pub mod decode;
 mod docs;
 pub mod get_block;
 pub mod get_blocks_head_header;
+pub mod processing;
 mod types;
 pub mod utils;
